refactor: reimagine strict node mode
