+++ packages/vite-node/package.json
@@ -78,6 +78,7 @@
   "dependencies": {
     "cac": "^6.7.14",
     "debug": "^4.3.4",
+    "local-pkg": "^0.4.2",
     "mlly": "^1.1.0",
     "pathe": "^1.1.0",
     "picocolors": "^1.0.0",
