+++ packages/vitest/src/node/index.ts
@@ -3,8 +3,8 @@ export { createVitest } from './create'
 export { VitestPlugin } from './plugins'
 export { startVitest } from './cli-api'
 
-export { VitestExecutor } from '../runtime/execute'
-export type { ExecuteOptions } from '../runtime/execute'
+export { VitestExecutor } from '../runtime/executors/vitest'
+export type { ExecuteOptions } from '../runtime/executors/vitest'
 
 export type { TestSequencer, TestSequencerConstructor } from './sequencers/types'
 export { BaseSequencer } from './sequencers/BaseSequencer'
