+++ packages/vitest/src/node/pool.ts
@@ -172,11 +172,11 @@ function createChannel(ctx: Vitest) {
         const r = await ctx.vitenode.transformRequest(id)
         return r?.map as RawSourceMap | undefined
       },
-      fetch(id, strict) {
-        return ctx.resolver.fetchModule(id, strict)
+      fetch(id, options) {
+        return ctx.vitenode.fetchModule(id, options)
       },
-      resolveId(id, importer, strict) {
-        return ctx.resolver.resolveId(id, importer, strict)
+      resolveId(id, importer) {
+        return ctx.vitenode.resolveId(id, importer)
       },
       onPathsCollected(paths) {
         ctx.state.collectPaths(paths)
